requirements.txt
seleniumbase==4.25.0
undetected-chromedriver==3.5.5
fake-useragent==1.5.0
requests==2.32.3
beautifulsoup4==4.12.3
selenium-wire==5.1.0
flask==3.0.3
stem==2.0.0
pysocks==1.7.1
numpy==1.26.4
opencv-python==4.9.0.80
tensorflow==2.15.0
web3==6.15.1
paramiko==3.3.1
cryptography==41.0.8
psutil==5.9.6